requirements.txt
pytest>=7.0
flake8>=6.0
PyYAML>=6.0
python-dotenv>=1.0
pydantic>=2.0
spacy>=3.0
pandas>=2.0
scikit-learn>=1.0
pdfplumber>=0.10.0
python-docx>=1.0.0
